+++ packages/calcite-components/src/components/input-time-zone/input-time-zone.tsx
@@ -45,7 +45,12 @@ import {
   FormComponent,
   HiddenFormInputSlot,
 } from "../../utils/form";
-import { createTimeZoneItems, getUserTimeZoneName, getUserTimeZoneOffset } from "./utils";
+import {
+  createTimeZoneItems,
+  findTimeZoneItemByProp,
+  getUserTimeZoneName,
+  getUserTimeZoneOffset,
+} from "./utils";
 import { InputTimeZoneMessages } from "./assets/input-time-zone/t9n";
 import { TimeZoneItem, TimeZoneMode } from "./interfaces";
 
@@ -71,6 +76,13 @@ export class InputTimeZone
   //  Properties
   //
   //--------------------------------------------------------------------------
+
+  /**
+   * When `true`, an empty value (`null`) will be allowed as a `value`.
+   *
+   * When `false`, an offset or name value is enforced, and clearing the input or blurring will restore the last valid `value`.
+   */
+  @Prop({ reflect: true }) clearable = false;
 
   /**
    * When `true`, interaction is prevented and the component is displayed with lower opacity.
@@ -183,6 +195,14 @@ export class InputTimeZone
 
   @Watch("value")
   handleValueChange(value: string, oldValue: string): void {
+    value = this.normalizeValue(value);
+
+    if (!value && this.clearable) {
+      this.value = value;
+      this.selectedTimeZoneItem = null;
+      return;
+    }
+
     const timeZoneItem = this.findTimeZoneItem(value);
 
     if (!timeZoneItem) {
@@ -297,10 +317,20 @@ export class InputTimeZone
   private onComboboxChange = (event: CustomEvent): void => {
     event.stopPropagation();
     const combobox = event.target as HTMLCalciteComboboxElement;
-    const selected = this.findTimeZoneItem(combobox.selectedItems[0].getAttribute("data-value"));
+    const selectedItem = combobox.selectedItems[0];
+
+    if (!selectedItem) {
+      this.value = null;
+      this.selectedTimeZoneItem = null;
+      this.calciteInputTimeZoneChange.emit();
+      return;
+    }
+
+    const selected = this.findTimeZoneItemByLabel(selectedItem.textLabel);
 
     const selectedValue = `${selected.value}`;
-    if (this.value === selectedValue) {
+
+    if (this.value === selectedValue && selected.label === this.selectedTimeZoneItem.label) {
       return;
     }
 
@@ -321,27 +351,27 @@ export class InputTimeZone
     this.calciteInputTimeZoneOpen.emit();
   };
 
-  private findTimeZoneItem(value: number | string): TimeZoneItem {
-    const valueToMatch = value;
+  private findTimeZoneItem(value: number | string | null): TimeZoneItem | null {
+    return findTimeZoneItemByProp(this.timeZoneItems, "value", value);
+  }
 
-    return this.timeZoneItems.find(
-      ({ value }) =>
-        // intentional == to match string to number
-        value == valueToMatch,
-    );
+  private findTimeZoneItemByLabel(label: string | null): TimeZoneItem | null {
+    return findTimeZoneItemByProp(this.timeZoneItems, "label", label);
   }
 
   private async updateTimeZoneItemsAndSelection(): Promise<void> {
     this.timeZoneItems = await this.createTimeZoneItems();
 
+    if (this.value === "" && this.clearable) {
+      this.selectedTimeZoneItem = null;
+      return;
+    }
+
     const fallbackValue = this.mode === "offset" ? getUserTimeZoneOffset() : getUserTimeZoneName();
     const valueToMatch = this.value ?? fallbackValue;
 
-    this.selectedTimeZoneItem = this.findTimeZoneItem(valueToMatch);
-
-    if (!this.selectedTimeZoneItem) {
-      this.selectedTimeZoneItem = this.findTimeZoneItem(fallbackValue);
-    }
+    this.selectedTimeZoneItem =
+      this.findTimeZoneItem(valueToMatch) || this.findTimeZoneItem(fallbackValue);
   }
 
   private async createTimeZoneItems(): Promise<TimeZoneItem[]> {
@@ -379,13 +409,18 @@ export class InputTimeZone
     disconnectMessages(this);
   }
 
+  private normalizeValue(value: string | null): string {
+    return value === null ? "" : value;
+  }
+
   async componentWillLoad(): Promise<void> {
     setUpLoadableComponent(this);
     await setUpMessages(this);
+    this.value = this.normalizeValue(this.value);
 
     await this.updateTimeZoneItemsAndSelection();
 
-    const selectedValue = `${this.selectedTimeZoneItem.value}`;
+    const selectedValue = this.selectedTimeZoneItem ? `${this.selectedTimeZoneItem.value}` : null;
     afterConnectDefaultValueSet(this, selectedValue);
     this.value = selectedValue;
   }
@@ -403,7 +438,7 @@ export class InputTimeZone
       <Host>
         <InteractiveContainer disabled={this.disabled}>
           <calcite-combobox
-            clearDisabled={true}
+            clearDisabled={!this.clearable}
             disabled={this.disabled}
             label={this.messages.chooseTimeZone}
             lang={this.effectiveLocale}
@@ -415,9 +450,12 @@ export class InputTimeZone
             onCalciteComboboxOpen={this.onComboboxOpen}
             open={this.open}
             overlayPositioning={this.overlayPositioning}
+            placeholder={
+              this.mode === "name" ? this.messages.namePlaceholder : this.messages.offsetPlaceholder
+            }
             readOnly={this.readOnly}
             scale={this.scale}
-            selectionMode="single-persist"
+            selectionMode={this.clearable ? "single" : "single-persist"}
             status={this.status}
             validation-icon={this.validationIcon}
             validation-message={this.validationMessage}